Add filename column to database table
Previously we only had pathnames so it was kind of
not conventient to send the entire path of server dirs
to the client.
So we instead created a new column called filename which will
have the basenames of the files. ex: Upload...jpg
So we can easily send them over via get method.